All requirements are completed

+++ SMS/Course.cs
@@ -11,6 +11,6 @@ namespace SMS
         public string CourseId { get; set; } = string.Empty;
         public string CourseName { get; set; } = string.Empty;
         public string InstructorName {  get; set; } = string.Empty;
-        public int NumberOfCredits { get; set; }
+        public double NumberOfCredits { get; set; }
     }
 }
